requirements.txt
transformers==4.8.1
justext==2.2.0
rank_bm25==0.2.1
torch==1.8.1
googlesearch-python==1.0.1
fastapi==0.67.0
uvicorn==0.14.0
